bump to v25.0.1

+++ include/autd3/def.hpp
@@ -15,13 +15,13 @@
 
 namespace autd3::driver {
 
-using Vector3 = Eigen::Matrix<double, 3, 1>;
-using Vector4 = Eigen::Matrix<double, 4, 1>;
-using Matrix4X4 = Eigen::Matrix<double, 4, 4>;
-using Matrix3X3 = Eigen::Matrix<double, 3, 3>;
-using Quaternion = Eigen::Quaternion<double>;
-using Affine3 = Eigen::Transform<double, 3, Eigen::Affine>;
+using Vector3 = Eigen::Matrix<float, 3, 1>;
+using Vector4 = Eigen::Matrix<float, 4, 1>;
+using Matrix4X4 = Eigen::Matrix<float, 4, 4>;
+using Matrix3X3 = Eigen::Matrix<float, 3, 3>;
+using Quaternion = Eigen::Quaternion<float>;
+using Affine3 = Eigen::Transform<float, 3, Eigen::Affine>;
 
-constexpr double pi = std::numbers::pi_v<double>;
+constexpr float pi = std::numbers::pi_v<float>;
 
 }  // namespace autd3::driver
