+++ include/autd3/driver/autd3_device.hpp
@@ -13,9 +13,9 @@ class AUTD3 {
   static constexpr size_t NUM_TRANS_IN_UNIT = native_methods::NUM_TRANS_IN_UNIT;
   static constexpr size_t NUM_TRANS_IN_X = native_methods::NUM_TRANS_IN_X;
   static constexpr size_t NUM_TRANS_IN_Y = native_methods::NUM_TRANS_IN_Y;
-  static constexpr double TRANS_SPACING = native_methods::TRANS_SPACING_MM;
-  static constexpr double DEVICE_WIDTH = native_methods::DEVICE_WIDTH_MM;
-  static constexpr double DEVICE_HEIGHT = native_methods::DEVICE_HEIGHT_MM;
+  static constexpr float TRANS_SPACING = native_methods::TRANS_SPACING_MM;
+  static constexpr float DEVICE_WIDTH = native_methods::DEVICE_WIDTH_MM;
+  static constexpr float DEVICE_HEIGHT = native_methods::DEVICE_HEIGHT_MM;
 
   AUTD3_API explicit AUTD3(Vector3 pos) : _position(std::move(pos)), _rotation(Quaternion::Identity()) {}
