+++ include/autd3/driver/firmware/fpga/phase.hpp
@@ -10,7 +10,7 @@ class Phase final {
   AUTD3_API explicit constexpr Phase(const uint8_t value) : _value(value) {}
   AUTD3_API explicit Phase(const Angle value) : _value(native_methods::AUTDPhaseFromRad(value.radian())) {}
 
-  AUTD3_API [[nodiscard]] double radian() const { return native_methods::AUTDPhaseToRad(_value); }
+  AUTD3_API [[nodiscard]] float radian() const { return native_methods::AUTDPhaseToRad(_value); }
 
   AUTD3_API [[nodiscard]] uint8_t value() const noexcept { return _value; }
